Try to read from DB

+++ AuctionApp/Core/Bid.cs
@@ -7,19 +7,19 @@ namespace AuctionApp.Core
         public int Id { get; set; }  
 
         public string BidderId { get; set; }  
-        public decimal Amount { get; set; }  
+        public double Amount { get; set; }  
         
         private DateTime _timePlaced;
         public DateTime TimePlaced => _timePlaced;  
         
-        public Bid(string bidderId, decimal amount)
+        public Bid(string bidderId, double amount)
         {
             BidderId = bidderId;
             Amount = amount;
             _timePlaced = DateTime.Now;  
         }
         
-        public Bid(int id, string bidderId, decimal amount, DateTime timePlaced)
+        public Bid(int id, string bidderId, double amount, DateTime timePlaced)
         {
             Id = id;
             BidderId = bidderId;
